+++ server/db/actions/Dog.js
@@ -2,9 +2,17 @@ import dbConnect from "../dbConnect";
 import Dog from "../models/Dog";
 
 export async function getDogs(filter = {}) {
-  await dbConnect();
-
   try {
+    await dbConnect();
+
+    if (filter["name"]) {
+      filter.name = new RegExp(filter.name, "i");
+    }
+
+    if (filter["instructors"]) {
+      filter.instructors = { $in: filter.instructors };
+    }
+
     return Dog.find(filter);
   } catch (e) {
     throw new Error("Unable to get dogs at this time, please try again");
@@ -12,9 +20,8 @@ export async function getDogs(filter = {}) {
 }
 
 export async function getDogById(id) {
-  await dbConnect();
-
   try {
+    await dbConnect();
     return Dog.findById(id);
   } catch (e) {
     throw new Error("Unable to get dog by Id at this time, please try again");
