+++ src/pages/api/dogs/index.js
@@ -3,14 +3,14 @@ import { getDogs } from "../../../../server/db/actions/Dog";
 export default async function handler(req, res) {
   if (req.method == "GET") {
     try {
-      const dogs = await getDogs();
+      const data = await getDogs();
 
       return res.status(200).json({
         success: true,
-        payload: dogs,
+        payload: data,
       });
     } catch (error) {
-      res.status(500).json({ error: error.message });
+      res.status(500).json({ success: false, message: error.message });
       return;
     }
   }
