+++ src/pages/api/users/index.js
@@ -2,6 +2,8 @@ import { createUser } from "../../../../server/db/actions/User";
 
 export default async function handler(req, res) {
   const { username, hash } = req.query;
+
+  console.log(hash);
   let userId;
   try {
     userId = await createUser(username, hash);
